+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # eventschopsolution Changelog
 
+## Version 1.0.0 (2022-06-25)
+
+* Add: Filter für Kategorien
+* Add: Filter für Start- und Endedatum
+* Add: Filter für Suchbegriff (Titel, Teaser)
+
 ## Version 0.0.2 (2022-06-24)
 
 * Fix: composer.json